[gym] Use float32 action space instead of float64 because it is the usual dtype for neural networks.

+++ python/gym_jiminy/common/gym_jiminy/common/envs/env_generic.py
@@ -391,12 +391,16 @@ class BaseJiminyEnv(ObserverControllerInterface, gym.Env):
                     effort_limit[motor.joint_velocity_idx] = \
                         MOTOR_EFFORT_MAX
 
-        # Set the action space
+        # Set the action space.
+        # Note that float32 is used instead of float64, because otherwise it
+        # would requires the neural network to perform float64 computations
+        # or cast the output for no really advantage since the action is
+        # directly forwarded to the motors, without intermediary computations.
         motors_velocity_idx = self.robot.motors_velocity_idx
         self.action_space = gym.spaces.Box(
-            low=-effort_limit[motors_velocity_idx],
-            high=effort_limit[motors_velocity_idx],
-            dtype=np.float64)
+            low=-effort_limit[motors_velocity_idx].astype(np.float32),
+            high=effort_limit[motors_velocity_idx].astype(np.float32),
+            dtype=np.float32)
 
     def reset(self,
               controller_hook: Optional[Callable[[], Optional[Tuple[
